refactored get odometer history tests
