+++ tests/unit/CertificateGenerationService.unitTest.ts
@@ -4,13 +4,6 @@ import 'reflect-metadata';
 const mockGetProfile = jest.fn();
 
 import Container from 'typedi';
-import sinon from 'sinon';
-import { CertificateGenerationService } from '../../src/services/CertificateGenerationService';
-import testResultsResp from '../resources/test-results-response.json';
-import testResultsRespFail from '../resources/test-results-fail-response.json';
-import testResultsRespPrs from '../resources/test-results-prs-response.json';
-import testResultsRespEmpty from '../resources/test-results-empty-response.json';
-import testResultsRespNoCert from '../resources/test-results-nocert-response.json';
 import { LambdaService } from '../../src/services/LambdaService';
 import { CERTIFICATE_DATA } from '../../src/models/Enums';
 import queueEventFail from '../resources/queue-event-fail.json';
@@ -18,9 +11,7 @@ import queueEventFailPRS from '../resources/queue-event-fail-prs.json';
 import { S3BucketService } from '../../src/services/S3BucketService';
 import { S3BucketMockService } from '../models/S3BucketMockService';
 import { LambdaMockService } from '../models/LambdaMockService';
-import { TechRecordsRepository } from '../../src/services/TechRecordsRepository';
 import { CertificatePayloadGenerator } from '../../src/services/CertificatePayloadGenerator';
-import { TestResultRepository } from '../../src/services/TestResultRepository';
 
 jest.mock('@dvsa/cvs-microservice-common/feature-flags/profiles/vtx', () => ({
   getProfile: mockGetProfile,
@@ -33,153 +24,8 @@ describe('Certificate Generation Service', () => {
   const invokeSpy = jest.spyOn(lambdaService, 'invoke');
   Container.set(LambdaService, lambdaService);
 
-  const techRecordsRepository = Container.get(TechRecordsRepository);
-  const searchTechRecordsSpy = jest.spyOn(techRecordsRepository, 'callSearchTechRecords');
-  const callGetTechRecordSpy = jest.spyOn(techRecordsRepository, 'callGetTechRecords');
-  Container.set(TechRecordsRepository, techRecordsRepository);
-
-  const testResultRepository = Container.get(TestResultRepository);
-  // const getOdometerSpy = jest.spyOn(testResultRepository, 'getOdometerHistory');
-  // Container.set(TestResultRepository, testResultRepository);
-
-  const sandbox = sinon.createSandbox();
-
   afterEach(() => {
-    sandbox.restore();
-    searchTechRecordsSpy.mockReset();
-    callGetTechRecordSpy.mockReset();
     invokeSpy.mockReset();
-  });
-
-  describe('getOdometerHistory function', () => {
-    context('when given a systemNumber with only failed test results', () => {
-      it('should return an empty odometer history list', async () => {
-        invokeSpy
-          .mockResolvedValueOnce(AWSResolve(JSON.stringify(testResultsRespFail)));
-
-        const systemNumberMock = '12345678';
-        const odometerHistory = await testResultRepository.getOdometerHistory(
-          systemNumberMock,
-        );
-
-        expect(invokeSpy).toBeCalledTimes(1);
-        expect(odometerHistory).toEqual({ OdometerHistoryList: [] });
-      });
-    });
-
-    context('when given a systemNumber which returns more than 3 pass or prs', () => {
-      it('should return an odometer history no greater than 3', async () => {
-        invokeSpy
-          .mockResolvedValue(AWSResolve(JSON.stringify(testResultsResp)));
-        const systemNumberMock = '12345678';
-        const odometerHistory = await testResultRepository.getOdometerHistory(
-          systemNumberMock,
-        );
-        expect(invokeSpy).toBeCalledTimes(1);
-        expect(odometerHistory).toEqual({
-          OdometerHistoryList: [
-            {
-              value: 400000,
-              unit: 'kilometres',
-              date: '19.01.2019',
-            },
-            {
-              value: 390000,
-              unit: 'kilometres',
-              date: '18.01.2019',
-            },
-            {
-              value: 380000,
-              unit: 'kilometres',
-              date: '17.01.2019',
-            },
-          ],
-        });
-      });
-    });
-
-    context('when given a systemNumber which returns tests which include those that are not Annual With Certificate', () => {
-      it('should omiting results that are not Annual With Certificate', async () => {
-        invokeSpy
-          .mockResolvedValue(AWSResolve(JSON.stringify(testResultsRespNoCert)));
-        const systemNumberMock = '12345678';
-        const odometerHistory = await testResultRepository.getOdometerHistory(
-          systemNumberMock,
-        );
-        expect(invokeSpy).toBeCalledTimes(1);
-        expect(odometerHistory).toEqual({
-          OdometerHistoryList: [
-            {
-              value: 400000,
-              unit: 'kilometres',
-              date: '19.01.2019',
-            },
-            {
-              value: 380000,
-              unit: 'kilometres',
-              date: '17.01.2019',
-            },
-            {
-              value: 360000,
-              unit: 'kilometres',
-              date: '15.01.2019',
-            },
-          ],
-        });
-      });
-    });
-
-    context('when given a systemNumber which returns a test result which was fail then prs', () => {
-      it('should return an odometer history which includes test result', async () => {
-        invokeSpy
-          .mockResolvedValue(AWSResolve(JSON.stringify(testResultsRespPrs)));
-        const systemNumberMock = '12345678';
-        const odometerHistory = await testResultRepository.getOdometerHistory(
-          systemNumberMock,
-        );
-        expect(invokeSpy).toBeCalledTimes(1);
-        expect(odometerHistory).toEqual({
-          OdometerHistoryList: [
-            {
-              value: 350000,
-              unit: 'kilometres',
-              date: '14.01.2019',
-            },
-          ],
-        });
-      });
-    });
-
-    context('when given a systemNumber which returns a test result which has no test types array', () => {
-      it('should omit the result from the odometer history', async () => {
-        invokeSpy
-          .mockResolvedValue(AWSResolve(JSON.stringify(testResultsRespEmpty)));
-        const systemNumberMock = '12345678';
-        const odometerHistory = await testResultRepository.getOdometerHistory(
-          systemNumberMock,
-        );
-        expect(invokeSpy).toBeCalledTimes(1);
-        expect(odometerHistory).toEqual({
-          OdometerHistoryList: [
-            {
-              value: 400000,
-              unit: 'kilometres',
-              date: '19.01.2019',
-            },
-            {
-              value: 380000,
-              unit: 'kilometres',
-              date: '17.01.2019',
-            },
-            {
-              value: 370000,
-              unit: 'kilometres',
-              date: '16.01.2019',
-            },
-          ],
-        });
-      });
-    });
   });
 
   const payloadGenerator = Container.get(CertificatePayloadGenerator);
